Fix: broken tests after refactor to columns directory
