+++ tests/unit/steps/columns/test_group.py
@@ -12,11 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any, Dict, List, Optional
 
-import pytest
 from distilabel.pipeline.local import Pipeline
-from distilabel.steps.columns.group import GroupColumns, MergeColumns
+from distilabel.steps.columns.group import GroupColumns
 
 
 class TestGroupColumns:
@@ -46,33 +44,3 @@ class TestGroupColumns:
         )
         output = next(combine.process([{"a": 1, "b": 2}], [{"a": 3, "b": 4}]))
         assert output == [{"merged_a": [1, 3], "merged_b": [2, 4]}]
-
-
-class TestMergeColumns:
-    @pytest.mark.parametrize(
-        "output_column, expected",
-        [
-            (None, "combined_column"),
-            ("queries", "queries"),
-        ],
-    )
-    def test_init(self, output_column: Optional[str], expected: str) -> None:
-        task = MergeColumns(columns=["query", "queries"], output_column=output_column)
-
-        assert task.inputs == ["query", "queries"]
-        assert task.outputs == [expected]
-
-    @pytest.mark.parametrize(
-        "columns",
-        [
-            [{"query": 1, "queries": 2}],
-            [{"query": 1, "queries": [2]}],
-            [{"query": [1], "queries": [2]}],
-        ],
-    )
-    def test_process(self, columns: List[Dict[str, Any]]) -> None:
-        combiner = MergeColumns(
-            columns=["query", "queries"],
-        )
-        output = next(combiner.process(columns))
-        assert output == [{"combined_column": [1, 2]}]
